Add testing for boolean arithmetics of interval membership
